Refactor web/main.py into modular directory structure
Split 2965-line monolithic main.py into proper package structure:

Structure:
- web/__init__.py: Package marker
- web/constants.py: Paths, security limits, regex patterns
- web/models.py: Pydantic models (Settings, Status, ApiResponse, etc.)
- web/services/: Business logic
  - daemon_client.py: ZeroMQ client for daemon communication
  - config.py: Configuration load/save
  - daemon.py: Daemon control (start/stop/status)
  - eq.py: EQ validation, sanitization, parsing
  - alsa.py: ALSA device discovery
- web/routers/: API endpoints
  - status.py: /, /status, /ws/stats, /settings, /devices
  - daemon.py: /daemon/* endpoints
  - eq.py: /eq/* endpoints
  - opra.py: /opra/* endpoints
- web/templates/: HTML templates
  - user.py: User-facing UI
  - admin.py: Admin dashboard

Benefits:
- main.py reduced from 2965 to ~55 lines
- Clear separation of concerns
- Easier testing and maintenance
- Proper Python package structure
