+++ web/__init__.py
@@ -1,0 +1,1 @@
+"""GPU Upsampler Web Control API package."""
